fisinhed test

+++ src/domain/users/use-cases/remove.avatar.ts
@@ -1,5 +1,7 @@
 import { Injectable } from '@nestjs/common';
+import Error400 from '~/helpers/errors/400.error';
 import Error404 from '~/helpers/errors/404.error';
+import { deleteFile } from '~/helpers/fileSystem';
 import { IUserRepository } from '~/infra/repository/user/IUserRepository';
 
 @Injectable()
@@ -13,6 +15,13 @@ export class RemoveAvatar {
       throw new Error404(`Could not find this user id ${extId}`);
     }
 
-    await this.userRepository.removeAvatar(extId);
+    if (!alreadyExists.avatar?.length) {
+      throw new Error400('This user dont have any avatar');
+    }
+
+    await Promise.all([
+      this.userRepository.removeAvatar(extId),
+      deleteFile(alreadyExists.avatar),
+    ]);
   }
 }
